+++ pkg/gen/filters/common/common_test.go
@@ -1,0 +1,1 @@
+package common
