+++ pkg/gen/filters/common/strings.go
@@ -1,0 +1,61 @@
+package common
+
+import (
+	"strings"
+)
+
+// UpperFirst returns a string representation of the value with the first character in UPPER CASE.
+func UpperFirst(s string) string {
+	if len(s) == 0 {
+		return s
+	}
+	return strings.ToUpper(s[:1]) + s[1:]
+}
+
+// LowerFirst returns a string representation of the value with the first character in lowercase.
+func LowerFirst(s string) string {
+	if len(s) == 0 {
+		return s
+	}
+	return strings.ToLower(s[:1]) + s[1:]
+}
+
+// FirstChar returns the first character of the value.
+func FirstChar(s string) string {
+	if len(s) == 0 {
+		return s
+	}
+	return s[:1]
+}
+
+// FirstCharLower returns the first character of the value in lowercase.
+func FirstCharLower(s string) string {
+	if len(s) == 0 {
+		return s
+	}
+	return strings.ToLower(s[:1])
+}
+
+// FirstCharUpper returns the first character of the value in UPPER CASE.
+func FirstCharUpper(s string) string {
+	if len(s) == 0 {
+		return s
+	}
+	return strings.ToUpper(s[:1])
+}
+
+func Join(sep string, a []string) string {
+	return strings.Join(a, sep)
+}
+
+func TrimPrefix(s, prefix string) string {
+	return strings.TrimPrefix(s, prefix)
+}
+
+func TrimSuffix(s, postfix string) string {
+	return strings.TrimSuffix(s, postfix)
+}
+
+func Replace(s, old, new string) string {
+	return strings.Replace(s, old, new, -1)
+}
